fix: stripped caffeine dependency.
tests currently failing, since they're not prepared for the call to be
async. But if we want it to work for Android we will need to perform
async calls. Otherwise we're accessing network on main thread

+++ src/main/kotlin/io/getunleash/UnleashConfig.kt
@@ -13,7 +13,7 @@ data class UnleashConfig(
     val url: String,
     val clientKey: String,
     val appName: String,
-    val refreshInterval: Int = 30,
-    val metricsInterval: Int = 30,
+    val refreshInterval: Long = 30,
+    val metricsInterval: Long = 30,
     val environment: String = "default"
 )
